Extract SearchTracks component into own folder
